Add build command
Added a command that can be used to build the project with.
The command uses Wepback to build the project with.

To run the command open up the terminal.
Make sure you're in the project repository.
Now copy and paste the follow into the terminal:

npm run build

Now hit enter. The main.js file in dist will be the final product.